Btrfs-progs: add UUID switches to mkfs and convert
Add the following switches to mkfs.btrfs and update man page:
* -U UUID, --uuid UUID

Add the following switches to btrfs-convert:
* -U UUID
* -U new
  Generates a random UUID (default behavior).
* -U copy
  Copies the UUID from the ext2fs.

Change-Id: I2bed896e4542b42374967d13047cbca6e3820179
Signed-off-by: Florian Albrechtskirchinger <falbrechtskirchinger@gmail.com>
Signed-off-by: William Douglas <william.douglas@intel.com>

---

Added for installer to be able to use UUID's passed from MIC. This
patch has not been accepted upstream but only feedback was adding
collision detection which we can risk not putting in here.

See https://patchwork.kernel.org/patch/1937711/ for further
discussion.